TODO_LIST is now updated AFTER all images have been downloaded without error.
Images are now stored in the 'pics' directory and all users symlink their pics to the 'pics' directory. This facilitates disk usage by not actually downloading files that we already have.

Added a mechanism to check if we are going to be downloading an infinite amount of pages.

Added some more runtime error checking so our script won't die in the middle of an error.

get() now returns None, None, None, allowing for the program to continue on a failed page or image get.